pyproject.toml
[tool.poetry]
name = "sep-2023-drf"
version = "0.1.0"
description = ""
authors = ["aisclass <aisclass@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
#django-cors-headers = "^4.3.1"
serializer = "^0.2.1"
serializers = "^0.2.4"
celery = "^5.4.0"
redis = "^5.0.4"
myapplication = "^0.1.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.6.0"
drf-yasg = "^1.21.7"
daphne = "^4.1.2"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.2.0"




[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
